- remove wallet manager fallback - refactor core/identity package
