+++ token/core/identity/msp/idemix/id.go
@@ -11,11 +11,10 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/hyperledger-labs/fabric-smart-client/platform/view/driver"
-
 	bccsp "github.com/IBM/idemix/bccsp/schemes"
 	csp "github.com/IBM/idemix/bccsp/schemes"
 	"github.com/golang/protobuf/proto"
+	"github.com/hyperledger-labs/fabric-smart-client/platform/view/driver"
 	m "github.com/hyperledger/fabric-protos-go/msp"
 	"github.com/hyperledger/fabric/msp"
 	"github.com/pkg/errors"
