+++ token/core/identity/msp/sig.go
@@ -4,7 +4,7 @@ Copyright IBM Corp. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package tms
+package msp
 
 import (
 	api2 "github.com/hyperledger-labs/fabric-smart-client/platform/fabric/driver"
