+++ token/core/identity/msp/x509/mspx509.go
@@ -11,10 +11,9 @@ import (
 
 	"github.com/golang/protobuf/proto"
 	"github.com/hyperledger-labs/fabric-smart-client/platform/view/view"
+	"github.com/hyperledger-labs/fabric-token-sdk/token/driver"
 	"github.com/hyperledger/fabric-protos-go/msp"
 	"github.com/pkg/errors"
-
-	"github.com/hyperledger-labs/fabric-token-sdk/token/driver"
 )
 
 // MSPIdentityDeserializer takes as MSP identity and returns an ECDSA verifier
