+++ token/core/zkatdlog/nogh/deserializer.go
@@ -10,15 +10,14 @@ import (
 	"bytes"
 	"sync"
 
-	"github.com/hyperledger-labs/fabric-smart-client/platform/view/view"
-	"github.com/pkg/errors"
-
 	idemix2 "github.com/hyperledger-labs/fabric-smart-client/platform/fabric/core/generic/msp/idemix"
+	"github.com/hyperledger-labs/fabric-smart-client/platform/view/view"
 	"github.com/hyperledger-labs/fabric-token-sdk/token/core/identity"
-	"github.com/hyperledger-labs/fabric-token-sdk/token/core/identity/fabric/idemix"
-	"github.com/hyperledger-labs/fabric-token-sdk/token/core/identity/fabric/x509"
+	"github.com/hyperledger-labs/fabric-token-sdk/token/core/identity/msp/idemix"
+	"github.com/hyperledger-labs/fabric-token-sdk/token/core/identity/msp/x509"
 	"github.com/hyperledger-labs/fabric-token-sdk/token/core/zkatdlog/crypto"
 	"github.com/hyperledger-labs/fabric-token-sdk/token/driver"
+	"github.com/pkg/errors"
 )
 
 type VerifierDES interface {
